feat: implement JoinCommunityController with create, store, and update methods; update routes and views accordingly
