+++ app/Http/Nomadelfia/Persona/Controllers/PersonController.php
@@ -6,9 +6,51 @@ namespace App\Nomadelfia\Persona\Controllers;
 
 use App\Nomadelfia\Persona\Models\Persona;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Http\Request;;
 
 final class PersonController
 {
+    public function create()
+    {
+        return view('nomadelfia.persone.anagrafica.create');
+    }
+
+    public function store(Request $request)
+    {
+        $request->validate([
+            'nominativo' => 'required',
+            'nome' => 'required',
+            'cognome' => 'required',
+            'data_nascita' => 'required|date',
+            'luogo_nascita' => 'required',
+            'sesso' => 'required',
+        ], [
+            'nominativo.required' => 'Il nominativo è obbligatorio',
+            'nominativo.unique' => 'IL nominativo inserito esiste già.',
+            'nome.required' => 'Il nome è obbligatorie',
+            'cognome.required' => 'Il cognome è obbligatorio',
+            'data_nascita.required' => 'La data di nascita è obbligatoria',
+            'luogo_nascita.required' => 'IL luogo di nascita è obbligatorio',
+            'sesso.required' => 'Il sesso della persona è obbligatorio',
+        ]);
+
+        // TODO: check the UNIQUE constraint on the persone table
+        $persona = Persona::create(
+            [
+                'nominativo' => $request->input('nominativo'),
+                'sesso' => $request->input('sesso'),
+                'nome' => $request->input('nome'),
+                'cognome' => $request->input('cognome'),
+                'provincia_nascita' => $request->input('luogo_nascita'),
+                'data_nascita' => $request->input('data_nascita'),
+                'id_arch_pietro' => 0,
+            ]
+        );
+        $persona->save();
+
+        return redirect(route('nomadelfia.persone.popolazine.entrata.create',  $persona->id))->withSuccess("Dati anagrafici di $persona->nominativo inseriti correttamente.");
+    }
+
     public function show($id)
     {
         $persona = Persona::findOrFail($id);
