+++ resources/views/nomadelfia/templates/entrataPersona.blade.php
@@ -1,7 +1,7 @@
 <form
     class="form"
     method="POST"
-    action="{{ route("nomadelfia.persone.popolazine.entrata.create", ["idPersona" => $persona->id]) }}"
+    action="{{ route("nomadelfia.persone.popolazine.entrata.create", ["id" => $persona->id]) }}"
 >
     @csrf
