+++ routes/web.php
@@ -38,7 +38,7 @@ use App\Nomadelfia\GruppoFamiliare\Controllers\PersonaGruppoFamiliareSpostaContr
 use App\Nomadelfia\Incarico\Controllers\IncarichiController;
 use App\Nomadelfia\Persona\Controllers\PersonIdentityController;
 use App\Nomadelfia\Persona\Controllers\PersonaDecessoController;
-use App\Nomadelfia\Persona\Controllers\PersonaEntrataController;
+use App\Nomadelfia\Persona\Controllers\JoinCommunityController;
 use App\Nomadelfia\Persona\Controllers\PersonaNominativoController;
 use App\Nomadelfia\Persona\Controllers\PersonaNumeroElencoController;
 use App\Nomadelfia\Persona\Controllers\PersonaPosizioneConcludiController;
@@ -106,16 +106,17 @@ Route::prefix('admin')->middleware('role:super-admin')->group(function () {
 
 Route::prefix('nomadelfia')->middleware('auth')->name('nomadelfia.')->group(function () {
     Route::get('/', [PopolazioneSummaryController::class, 'index'])->middleware('can:popolazione.persona.visualizza')->name('index');
+
+    Route::get('people', [PersonController::class, 'create'])->name('person.create');
+    Route::post('people', [PersonController::class, 'store'])->name('person.store');
     Route::get('people/{id}', [PersonController::class, 'show'])->middleware('can:popolazione.persona.visualizza')->name('person.show');
 
-    Route::get('people/identity', [PersonIdentityController::class, 'create'])->name('person.identity.create');
-    Route::post('people/identity', [PersonIdentityController::class, 'store'])->name('person.identity.store');
     Route::get('people/{id}/identity', [PersonIdentityController::class, 'edit'])->name('person.identity.edit');
-    Route::put('people/{id}/identtity', [PersonIdentityController::class, 'update'])->name('person.identity.update');
+    Route::put('people/{id}/identity', [PersonIdentityController::class, 'update'])->name('person.identity.update');
 
-    Route::get('persone/{idPersona}/entrata', [PersonaEntrataController::class, 'create'])->name('persone.popolazine.entrata.create');
-    Route::post('persone/{idPersona}/entrata', [PersonaEntrataController::class, 'store'])->name('persone.anagrafica.entrata.scelta');
-    Route::post('persone/{idPersona}/entrata/{entrata}/modifica', [PersonaEntrataController::class, 'update'])->name('persone.dataentrata.modifica');
+    Route::get('people/{id}/join', [JoinCommunityController::class, 'create'])->name('persone.popolazine.entrata.create');
+    Route::post('people/{id}/join', [JoinCommunityController::class, 'store'])->name('persone.anagrafica.entrata.scelta');
+    Route::post('people/{id}/join/{entrata}/modifica', [JoinCommunityController::class, 'update'])->name('persone.dataentrata.modifica');
 
     Route::post('persone/{idPersona}/decesso', [PersonaDecessoController::class, 'store'])->name('persone.decesso');
     Route::post('persone/{idPersona}/uscita', [PersonaUscitaController::class, 'store'])->name('persone.uscita');
