+++ tests/Popolazione/Http/PersonaAnagraficaTest.php
@@ -4,12 +4,13 @@ declare(strict_types=1);
 
 namespace Tests\Http\Nomadelfia;
 
+use App\Nomadelfia\Persona\Controllers\PersonController;
 use App\Nomadelfia\Persona\Controllers\PersonIdentityController;
 use App\Nomadelfia\Persona\Models\Persona;
 
 it('shows form to create persona', function (): void {
     login();
-    $this->get(action([PersonIdentityController::class, 'create']))
+    $this->get(action([PersonController::class, 'create']))
         ->assertSuccessful();
 });
 
@@ -31,7 +32,7 @@ it('can update anagrafica', function (): void {
     $newLuogo = 'my-luogo';
     $newSesso = 'F';
     $newbiografia = 'Sono nato e morto';
-    $this->put(action([PersonIdentityController::class, 'update'],$persona->id),
+    $this->put(action([PersonIdentityController::class, 'update'], $persona->id),
         [
             'nome' => $newName,
             'cognome' => $newSurname,
@@ -55,7 +56,7 @@ it('can update anagrafica', function (): void {
 it('can insert a persona', function (): void {
     login();
     $this->withoutExceptionHandling();
-    $this->post(action([PersonIdentityController::class, 'store']),
+    $this->post(action([PersonController::class, 'store']),
         [
             'nominativo' => 'my-name',
             'nome' => 'name',
