+++ tests/Popolazione/Http/PersonaControllerTest.php
@@ -10,7 +10,7 @@ use App\Nomadelfia\Persona\Models\Persona;
 
 it('shows form to insert a persona', function (): void {
     login();
-    $this->get(action([PersonIdentityController::class, 'create']))
+    $this->get(action([PersonController::class, 'create']))
         ->assertSuccessful();
 });
 
@@ -19,7 +19,7 @@ it('render the show of a person with minimal info', function (): void {
 
     login();
 
-    $this->get(action([PersonController::class, 'show'], ['id' => $persona->id]))
+    $this->get(action([PersonController::class, 'show'], $persona->id))
         ->assertOk()
         ->assertSeeText($persona->nominativo);
 });
